Add support for running tests through Molecule
This commit introduces the 'agent' test scenario that can be run with the
following command:
```
molecule test -s agent
```
Before the tests can be run, the instance type, region, and vpc information in
molecule/agent/molecule.yml needs to be populated. The GH Actions added in a
future commit will handle that automatically for CI runs.