Only install build deps when requested
By default, do not install the build dependencies for the kernel module or eBPF
probes. This places an implicit requirement that there must exist a prebuilt
probe available for download. The major benefit of this change is that we the
Role does not become responsible for any package management outside the Agent
itself, and the runtime is much faster.