fix: correct use of rpm as a comparison against ansible_pkg_mgr
main.yml had erroneously used 'rpm' in the list of stings to compare against
the value of ansible_pkg_mgr to determine which repository type to configure.
the issue is that the value that should be there is 'yum'. this regression was
introduced as a part of adding the debian package support in pull request #12.