fix: reduce complexities of kernel devel detections for EL distros
The `Install Kernel Headers (7/8)` steps in the AlmaLinux, CentOSLinux,
and RockyLinux dependency tasks utilized `vars` to loop over the supported
kernel types to determine the required kernel-devel package name. Though
running `ansible-lint` it was found that the way we were using `vars`, while
working, was failing schema validations as the `vars` field expects an object
and not a list like we were providing. This change removes the complexities
created by that logic and will instead result in needing to create a dedicated
file for Oracle Linux dependencies if we choose to add support for that
distribution in the Role in the future.