feat: bring the placeholder template variables in line with github secrets
The obvious placeholder variable names in the molecule.yml file in the `agent`
scenario were not compliant with GitHub Secrets. This change updates the
variable names so that GitHub Secrets can be used in the CI runs.

This change is going in before the actual workflow files themselves to
facilitate using `act` to test the workflows during development.