feat: create GitHub auto-release notes configuration
Utilize GitHub's ability to automatically generate release notes by providing
a configuration. The release notes will be generated by using labels placed on
PRs. The labels that are parsed are:
- breaking-change
- feature
- bug
- documentation

PRs tagged with `documentation` will be ignored, but each of the other labels
will have their own category in the Release Notes.