feat: allow custom package repositories
Users may wish to store the official packages on a custom repository internal
to their environment similar to how users often store the Agent container
images on an internal registry. This change allows for that functionality
by allowing custom paths for `rpm` and `deb` repository URLs as well as
GPG keys.