feat: allow installing the latest version of the agent package
The `agent_install` Role had gone out of its way to ensure that it was only
possible to install a specific (pinned) version of the Agent. This choice was
made to guarantee the idempotence of the Role. This unfortunately ignored the
set of users who either want to use the Role to install whatever the latest
version of the Agent is at the time, or simply don't mind if the Role is not
idempotent.

This change retains the existing behavior of having the version of the Agent
be pinned by default, but allows for the installation of the current latest
version of the Agent by either specifying an agent version of `latest` or ~.