fix: serialize updatecli runs
Running `updatecli` as a matrix is preferred because it simplifies the management of
needing to update several families of OS types in a common, but separate manner.
Unfortunately there appears to be an odd behavior with parallel executions of the
updatecli tool that results in each parallel run of the matrix using the same git branch
name, which results in a combination of either experienceing git fast-forward errors, or
more than one OS family's worth of updates on the same PR. This change restricts the matrix
to only being able to run serially until a more optimal solution can be found.