doc: clarify the release process adr
the release process, while defined in the adr, was not entirely clear
and lacked finalized diagrams in some cases. this change brings consistency
and clarity to the release process, as well as addresses how to handle the
adjustments of the Collection version as defined in the `galaxy.yaml`
metadata file.