fix(ebpf): robustify probe unloading in case of ebpf
there might be situations which we have a
downloaded kernel module loaded in the system.
In that case, modprobe -r will fail since the module
is not known to the system.
Also, the service unit should normally remove
it autonomously, so the explicit removal should
only be taken as the last resort.

So put in place the two following changes:
1. After stopping the unit, check if the
   module is still present -- do nothing if not
2. Try with modprobe -r first and if that fails,
   try with rmmod