feat(agent_install): only install required packages [SMAGENT-6558]
rework the role so to only install the packages required by the
selected driver (and therefore, the respective dependencies).
Also, remove packages which are no longer required by the
selected driver type.

Notice we need to craft a list augmented with the selected
version, if one is present.

Most of the selection logic (namely, based on the version)
resides in the 'agent' filter which is probably far from
ideal but gets the job done.